requirements.txt
# Web Framework e Servidor ASGI
fastapi==0.116.1
uvicorn[standard]==0.35.0

# Validação de Dados
pydantic==2.11.7
pydantic-settings==2.3.4
pydantic[email]==2.11.7

# Banco de Dados (PostgreSQL Async)
sqlalchemy[asyncio]==2.0.43
alembic==1.16.4
asyncpg==0.30.0

# Mensageria (Kafka Async)
aiokafka==0.12.0

# Cache (Redis Async)
redis==6.4.0

# Autenticação e Segurança (JWT e Hashing de Senhas)
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.3.0

# Utilitários
python-dotenv==1.1.1
python-multipart==0.0.9